requirements.txt
# Core data processing
pandas==2.2.1
numpy==1.26.4
scipy>=1.7.0

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pymysql>=1.0.0

# Machine Learning
scikit-learn==1.4.1
tensorflow>=2.8.0
prophet==1.1.5
statsmodels>=0.13.0

# Visualization
plotly==5.19.0
matplotlib>=3.5.0
seaborn>=0.11.0
altair>=4.2.0

# Web Application
streamlit==1.32.0
flask>=2.0.0
dash>=2.0.0

# Utilities
python-dotenv==1.0.1
tqdm>=4.62.0
joblib>=1.1.0
pytest>=6.2.5

# Data Processing
openpyxl>=3.0.0
xlrd>=2.0.0
pyarrow>=6.0.0

# API and Integration
requests>=2.26.0
beautifulsoup4>=4.9.0
pytz>=2021.3 